+++ src/main/java/com/joucode/campus_x_backend/user/infrastructure/adapters/output/persistence/entity/RoleEntity.java
@@ -1,6 +1,6 @@
 package com.joucode.campus_x_backend.user.infrastructure.adapters.output.persistence.entity;
 
-import com.joucode.campus_x_backend.user.domain.models.enums.RoleName;
+import com.joucode.campus_x_backend.user.domain.enums.RoleName;
 import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
@@ -17,7 +17,7 @@ public class RoleEntity implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "role_id")
-    private Integer roleId;
+    private Long roleId;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "name",length = 20, nullable = false)
